+++ Sources/Modeler/Camera.hpp
@@ -20,8 +20,8 @@ public:
 	const glm::dvec3&	GetCenter () const;
 	const glm::dvec3&	GetUp () const;
 
-	glm::mat4			GetViewMatrix () const;
-	glm::mat4			GetProjectionMatrix (double width, double height) const;
+	glm::dmat4			GetViewMatrix () const;
+	glm::dmat4			GetProjectionMatrix (double width, double height) const;
 
 	void				Orbit (double angleX, double angleY);
 	void				Pan (double diffX, double diffY);
